Add a user specified callback parameter to GetNamedDSMSegment
This function previously had a callback, but there was no way to pass
user data to the callback.

For example, the documentation mentions that when using
GetNamedDSMSegment, LWLocks should be initialized in the init callback.
But if the calling code is also generic, for example receiving a
tranche_name as a parameter, there's no good way to forward this
parameter inside the initialization callback.

This change solves the above problem with the generic callback
parameter.